requirements-dev.txt
-r requirements.txt
black==18.6b4  # format code
django-debug-toolbar==1.10.1  # debug toolbar on the web page
flake8==3.5.0
flake8-blind-except==0.1.1  # wide try except
flake8-bugbear==18.8.0  # search for possible bugs
flake8-comprehensions==1.4.1  # find where is better to use comprehensions
flake8-isort==2.5  # check  if imports are sorted
flake8-mutable==1.2.0  # check mutable imports
isort==4.3.4  # sort imports and seperate by each line
pytest-django>=3.3.0
pytest-mock>=1.10.0
